requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
ansicon=1.89.0=pypi_0
anyio=3.7.1=pypi_0
appdirs=1.4.4=pypi_0
astor=0.8.1=pypi_0
asttokens=2.4.1=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
beautifulsoup4=4.12.2=pypi_0
blessed=1.20.0=pypi_0
blinker=1.7.0=pypi_0
ca-certificates=2023.12.12=haa95532_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
chromedriver=2.24.1=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.0=pypi_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
distro=1.9.0=pypi_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pypi_0
fastapi=0.104.1=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.47.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.12.2=pypi_0
git-python=1.0.3=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
h11=0.14.0=pypi_0
html2image=2.0.4.3=pypi_0
httpcore=1.0.2=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.20.1=pypi_0
idna=3.6=pypi_0
importlib-metadata=6.11.0=pypi_0
importlib-resources=6.1.1=pypi_0
inquirer=3.1.4=pypi_0
ipykernel=6.28.0=pypi_0
ipython=8.18.1=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pypi_0
jinxed=1.2.1=pypi_0
jsonschema=4.20.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-core=5.6.1=pypi_0
kiwisolver=1.4.5=pypi_0
litellm=1.16.8=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
multidict=6.0.4=pypi_0
nest-asyncio=1.5.8=pypi_0
numpy=1.26.2=pypi_0
open-interpreter=0.1.18=pypi_0
openai=1.6.1=pypi_0
openssl=1.1.1w=h2bbff1b_0
orjson=3.9.10=pypi_0
outcome=1.3.0.post0=pypi_0
packaging=23.2=pypi_0
pandas=2.1.4=pypi_0
parso=0.8.3=pypi_0
pillow=10.1.0=pypi_0
pip=23.3.1=py39haa95532_0
platformdirs=4.1.0=pypi_0
prompt-toolkit=3.0.43=pypi_0
protobuf=4.25.1=pypi_0
psutil=5.9.7=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=14.0.2=pypi_0
pycparser=2.21=pypi_0
pydantic=2.5.3=pypi_0
pydantic-core=2.14.6=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
pyreadline3=3.4.1=pypi_0
pysocks=1.7.1=pypi_0
python=3.9.0=h6244533_2
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
python-editor=1.0.4=pypi_0
python-multipart=0.0.6=pypi_0
pytz=2023.3.post1=pypi_0
pywin32=306=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=pypi_0
readchar=4.0.5=pypi_0
redis=5.0.1=pypi_0
referencing=0.32.0=pypi_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.16.2=pypi_0
rq=1.15.1=pypi_0
selenium=4.16.0=pypi_0
setuptools=68.2.2=py39haa95532_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack-data=0.6.3=pypi_0
starlette=0.27.0=pypi_0
streamlit=1.29.0=pypi_0
tenacity=8.2.3=pypi_0
tiktoken=0.4.0=pypi_0
tokenizers=0.15.0=pypi_0
tokentrim=0.1.13=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.0=pypi_0
trio=0.23.2=pypi_0
trio-websocket=0.11.1=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2023.4=pypi_0
tzlocal=5.2=pypi_0
urllib3=2.1.0=pypi_0
uvicorn=0.22.0=pypi_0
validators=0.22.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=3.0.0=pypi_0
wcwidth=0.2.12=pypi_0
webdriver-manager=4.0.1=pypi_0
websocket-client=1.7.0=pypi_0
wget=3.2=pypi_0
wheel=0.41.2=py39haa95532_0
wsproto=1.2.0=pypi_0
yarl=1.9.4=pypi_0
zipp=3.17.0=pypi_0
